# postman/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Klar API",
    "description": "German bureaucratic mail \u2192 structured obligations + deadlines.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/actions": {
      "get": {
        "tags": [
          "actions"
        ],
        "summary": "List the current user's action items across all letters",
        "description": "Returns every action linked to a letter the user owns. Optional `?status=open|done|ignored` filter. Empty string is treated as 'no filter' so the frontend can bind directly to React state.",
        "operationId": "list_actions_api_actions_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Not authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unknown status value.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/actions/{action_id}": {
      "patch": {
        "tags": [
          "actions"
        ],
        "summary": "Update fields on an action (status, deadline, title, description)",
        "description": "Every field change is logged to the `UserCorrection` table \u2014 the feedback loop for future prompt tuning. Send only the fields you want to change.",
        "operationId": "update_action_api_actions__action_id__patch",
        "parameters": [
          {
            "name": "action_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Action Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Not authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "`ACTION_NOT_FOUND`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActionStatus": {
        "type": "string",
        "enum": [
          "open",
          "done",
          "ignored"
        ],
        "title": "ActionStatus"
      },
      "ActionUpdate": {
        "properties": {
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ActionStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "deadline": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "title": "ActionUpdate"
      },
...
      "ErrorDetails": {
        "properties": {
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ValidationFieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          },
          "declared": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Declared"
          },
          "detected": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detected"
          }
        },
        "type": "object",
        "title": "ErrorDetails",
        "description": "Optional structured extras attached to an error envelope."
      },
      "ErrorResponse": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Stable machine-readable identifier \u2014 see docs/06-api-contract.md."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Localized, user-facing copy."
          },
          "details": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetails"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorResponse",
        "description": "Every non-2xx response from this API has this shape.\n\nFrontend code should switch on `code` (stable identifier), display\n`message` to the user, and inspect `details` for field-level\ninformation on `VALIDATION_ERROR`s."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "ValidationFieldError": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field",
            "description": "Dotted path to the offending field."
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "field",
          "message"
        ],
        "title": "ValidationFieldError"
      }
    }
  }
}